+++ include/c/acosl.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-__ELIBC_CONST double acosl(double x) __ELIBC_NOTHROW;
+__ELIBC_CONST long double acosl(long double x) __ELIBC_NOTHROW;
 
 #ifdef __cplusplus
 }
