+++ include/c/erf.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-__ELIBC_CONST float erf(float x) __ELIBC_NOTHROW;
+__ELIBC_CONST double erf(double x) __ELIBC_NOTHROW;
 
 #ifdef __cplusplus
 }
