+++ include/c/ilogbf.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-__ELIBC_CONST float ilogbf(float x) __ELIBC_NOTHROW;
+__ELIBC_CONST int ilogbf(float x) __ELIBC_NOTHROW;
 
 #ifdef __cplusplus
 }
