+++ include/c/llround.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-__ELIBC_CONST double llround(double x) __ELIBC_NOTHROW;
+__ELIBC_CONST long long int llround(double x) __ELIBC_NOTHROW;
 
 #ifdef __cplusplus
 }
