+++ include/c/llroundf.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-__ELIBC_CONST float llroundf(float x) __ELIBC_NOTHROW;
+__ELIBC_CONST long long int llroundf(float x) __ELIBC_NOTHROW;
 
 #ifdef __cplusplus
 }
