+++ include/c/lrintf.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-__ELIBC_CONST float lrintf(float x) __ELIBC_NOTHROW;
+__ELIBC_CONST long int lrintf(float x) __ELIBC_NOTHROW;
 
 #ifdef __cplusplus
 }
