+++ src/math/llroundl.c
@@ -6,6 +6,6 @@
 /*------------------------------------------------------------------------------
 // Name: llroundl
 //----------------------------------------------------------------------------*/
-long double llroundl(long double x) {
+long long int llroundl(long double x) {
 	return __builtin_llroundl(x);
 }
